% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus-utils.R
\name{buildConsensusMatrix}
\alias{buildConsensusMatrix}
\title{Build consensus matrix showing co-clustering frequency}
\usage{
buildConsensusMatrix(valid_indices, cluster_assignments, n_total_spectra)
}
\arguments{
\item{valid_indices}{Indices of spectra that passed filtering}

\item{cluster_assignments}{Cluster assignments for filtered spectra}

\item{n_total_spectra}{Total number of spectra before filtering}
}
\value{
Matrix showing co-clustering frequency
}
\description{
Build consensus matrix showing co-clustering frequency
}
\keyword{internal}
