% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus-utils.R
\name{calculateCopheneticCorrelation}
\alias{calculateCopheneticCorrelation}
\title{Calculate cophenetic correlation for clustering quality assessment}
\usage{
calculateCopheneticCorrelation(consensus_matrix)
}
\arguments{
\item{consensus_matrix}{Consensus matrix showing co-clustering frequencies}
}
\value{
Cophenetic correlation coefficient
}
\description{
Calculate cophenetic correlation for clustering quality assessment
}
\keyword{internal}
