% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus-utils.R
\name{calculateStabilityMetrics}
\alias{calculateStabilityMetrics}
\title{Calculate stability metrics for consensus NMF using clustering approach}
\usage{
calculateStabilityMetrics(nmf_runs, consensus_result)
}
\arguments{
\item{nmf_runs}{List of NMF results}

\item{consensus_result}{List with consensus matrices and clustering info}
}
\value{
List with stability metrics
}
\description{
Calculate stability metrics for consensus NMF using clustering approach
}
\keyword{internal}
