% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAnalysis.R
\name{convertGeneIDs}
\alias{convertGeneIDs}
\title{Convert gene IDs to Entrez IDs}
\usage{
convertGeneIDs(genes, orgdb, from_type)
}
\arguments{
\item{genes}{Character vector of gene identifiers}

\item{orgdb}{Character, organism database name}

\item{from_type}{Character, source ID type}
}
\value{
Character vector of Entrez IDs
}
\description{
Convert gene IDs to Entrez IDs
}
\keyword{internal}
