% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus-utils.R
\name{cosine_distance}
\alias{cosine_distance}
\title{Calculate cosine distance between vectors}
\usage{
cosine_distance(x, y)
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}
}
\value{
Cosine distance (1 - cosine similarity)
}
\description{
Calculate cosine distance between vectors
}
\keyword{internal}
