% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAnalysis.R
\name{detectGeneIDType}
\alias{detectGeneIDType}
\title{Detect gene ID type automatically}
\usage{
detectGeneIDType(genes, orgdb)
}
\arguments{
\item{genes}{Character vector of gene identifiers}

\item{orgdb}{Character, organism database name}
}
\value{
Character, detected key type
}
\description{
Detect gene ID type automatically
}
\keyword{internal}
