% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus-utils.R
\name{findBestMatching}
\alias{findBestMatching}
\title{Find best matching between two factor matrices}
\usage{
findBestMatching(mat1, mat2)
}
\arguments{
\item{mat1}{First matrix}

\item{mat2}{Second matrix}
}
\value{
List with permutation and correlation matrix
}
\description{
Find best matching between two factor matrices
}
\keyword{internal}
