% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNMF-accessors.R
\name{getAvailableK}
\alias{getAvailableK}
\title{Get available k values from cNMF results}
\usage{
getAvailableK(x, name = "cNMF")
}
\arguments{
\item{x}{A SingleCellExperiment object with cNMF results}

\item{name}{Character, name of the cNMF result (default "cNMF")}
}
\value{
Integer vector of available k values
}
\description{
Get available k values from cNMF results
}
\examples{
# sce <- runConsensusNMF(sce, k_range = 5:10)
# k_values <- getAvailableK(sce)
}
