% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getCoefficients}
\alias{getCoefficients}
\title{Extract NMF coefficient matrix from SingleCellExperiment object}
\usage{
getCoefficients(x, name = "NMF")
}
\arguments{
\item{x}{A SingleCellExperiment object with NMF results}

\item{name}{Character, name of the NMF result to extract (default "NMF")}
}
\value{
Matrix with cells x factors (coefficient matrix, same as reducedDim)
}
\description{
Extract NMF coefficient matrix from SingleCellExperiment object
}
\examples{
# sce <- runNMF(sce, k = 10)
# coeffs <- getCoefficients(sce)
}
