% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNMF-accessors.R
\name{getConsensusGEPs}
\alias{getConsensusGEPs}
\title{Extract consensus gene expression programs (GEPs) from cNMF results}
\usage{
getConsensusGEPs(x, name = "cNMF", k = NULL)
}
\arguments{
\item{x}{A SingleCellExperiment object with cNMF results}

\item{name}{Character, name of the cNMF result to extract (default "cNMF")}

\item{k}{Integer, which k value to extract. If NULL, uses optimal k}
}
\value{
Matrix with features x programs (consensus basis matrix)
}
\description{
Extract consensus gene expression programs (GEPs) from cNMF results
}
\examples{
# sce <- runConsensusNMF(sce, k_range = 5:10)
# geps <- getConsensusGEPs(sce)
}
