% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNMF-accessors.R
\name{getConsensusNMFInfo}
\alias{getConsensusNMFInfo}
\title{Get consensus NMF information and parameters}
\usage{
getConsensusNMFInfo(x, name = "cNMF")
}
\arguments{
\item{x}{A SingleCellExperiment object with cNMF results}

\item{name}{Character, name of the cNMF result (default "cNMF")}
}
\value{
List with cNMF run information (excluding large matrices)
}
\description{
Get consensus NMF information and parameters
}
\examples{
# sce <- runConsensusNMF(sce, k_range = 5:10)
# info <- getConsensusNMFInfo(sce)
}
