% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNMF-accessors.R
\name{getOptimalK}
\alias{getOptimalK}
\title{Get optimal k value from cNMF results}
\usage{
getOptimalK(x, name = "cNMF")
}
\arguments{
\item{x}{A SingleCellExperiment object with cNMF results}

\item{name}{Character, name of the cNMF result (default "cNMF")}
}
\value{
Integer, optimal k value
}
\description{
Get optimal k value from cNMF results
}
\examples{
# sce <- runConsensusNMF(sce, k_range = 5:10)
# k_opt <- getOptimalK(sce)
}
