% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNMF-accessors.R
\name{getTopGEPFeatures}
\alias{getTopGEPFeatures}
\title{Get top features for each consensus gene expression program}
\usage{
getTopGEPFeatures(x, name = "cNMF", k = NULL, n = 20)
}
\arguments{
\item{x}{A SingleCellExperiment object with cNMF results}

\item{name}{Character, name of the cNMF result (default "cNMF")}

\item{k}{Integer, which k value to use. If NULL, uses optimal k}

\item{n}{Integer, number of top features per program (default 20)}
}
\value{
List of character vectors, top features for each GEP
}
\description{
Get top features for each consensus gene expression program
}
\examples{
# sce <- runConsensusNMF(sce, k_range = 5:10)
# top_genes <- getTopGEPFeatures(sce, n = 20)
}
