% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus-utils.R
\name{greedyMatching}
\alias{greedyMatching}
\title{Greedy matching for larger k}
\usage{
greedyMatching(cor_mat)
}
\arguments{
\item{cor_mat}{Correlation matrix}
}
\value{
Best permutation vector
}
\description{
Greedy matching for larger k
}
\keyword{internal}
