% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus-utils.R
\name{permutations}
\alias{permutations}
\title{Generate all permutations for small k}
\usage{
permutations(k)
}
\arguments{
\item{k}{Integer, number of elements}
}
\value{
Matrix of permutations
}
\description{
Generate all permutations for small k
}
\keyword{internal}
