% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNMF-visualization.R
\name{plotGEPs}
\alias{plotGEPs}
\title{Plot gene expression programs (GEPs) heatmap}
\usage{
plotGEPs(x, name = "cNMF", k = NULL, programs = NULL, n_genes = 20)
}
\arguments{
\item{x}{A SingleCellExperiment object with cNMF results}

\item{name}{Character, name of the cNMF result (default "cNMF")}

\item{k}{Integer, which k value to plot. If NULL, uses optimal k}

\item{programs}{Integer vector, which programs to plot (default all)}

\item{n_genes}{Integer, number of top genes per program to show (default 20)}
}
\value{
ggplot2 object
}
\description{
Plot gene expression programs (GEPs) heatmap
}
\examples{
# sce <- runConsensusNMF(sce, k_range = 5:10)
# plotGEPs(sce, programs = 1:3, n_genes = 15)
}
