% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNMF-visualization.R
\name{plotGEPUsage}
\alias{plotGEPUsage}
\title{Plot GEP usage in reduced dimension space}
\usage{
plotGEPUsage(x, name = "cNMF", k = NULL, programs = 1:4, reduction = "UMAP")
}
\arguments{
\item{x}{A SingleCellExperiment object with cNMF results}

\item{name}{Character, name of the cNMF result (default "cNMF")}

\item{k}{Integer, which k value to use. If NULL, uses optimal k}

\item{programs}{Integer vector, which programs to plot (default 1:4)}

\item{reduction}{Character, which reducedDim to use for coordinates (default "UMAP")}
}
\value{
ggplot2 object
}
\description{
Plot GEP usage in reduced dimension space
}
\examples{
# sce <- runConsensusNMF(sce, k_range = 5:10)
# sce <- runUMAP(sce)
# plotGEPUsage(sce, programs = 1:4, reduction = "UMAP")
}
