% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNMF-visualization.R
\name{plotStability}
\alias{plotStability}
\title{Plot stability metrics from consensus NMF}
\usage{
plotStability(
  x,
  name = "cNMF",
  metrics = c("stability", "silhouette", "reproducibility")
)
}
\arguments{
\item{x}{A SingleCellExperiment object with cNMF results}

\item{name}{Character, name of the cNMF result (default "cNMF")}

\item{metrics}{Character vector, which metrics to plot (default all)}
}
\value{
ggplot2 object
}
\description{
Plot stability metrics from consensus NMF
}
\examples{
# sce <- runConsensusNMF(sce, k_range = 5:10)
# plotStability(sce)
}
