% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goPlotting.R
\name{saveGOPlotsToPDF}
\alias{saveGOPlotsToPDF}
\title{Save GO Plots to PDF}
\usage{
saveGOPlotsToPDF(
  go_results_list,
  filename = "go_enrichment_plots.pdf",
  width = 12,
  height = 8
)
}
\arguments{
\item{go_results_list}{Named list of GO enrichment results}

\item{filename}{Output PDF filename}

\item{width}{PDF width in inches}

\item{height}{PDF height in inches}
}
\description{
Convenience function to save multiple GO plots to a single PDF
}
