% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus-utils.R
\name{solveNNLS}
\alias{solveNNLS}
\title{Solve non-negative least squares problem}
\usage{
solveNNLS(A, b)
}
\arguments{
\item{A}{Matrix A (features x factors)}

\item{b}{Vector b (features)}
}
\value{
Non-negative solution vector x
}
\description{
Solves min ||Ax - b||^2 subject to x >= 0
}
\keyword{internal}
