% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocNMF-package.R
\docType{package}
\name{BiocNMF-package}
\alias{BiocNMF}
\alias{BiocNMF-package}
\title{BiocNMF: Fast Non-negative Matrix Factorization for Single Cell and Spatial Data}
\description{
This package provides high-performance non-negative matrix factorization (NMF) methods for SingleCellExperiment and SpatialExperiment objects using the RcppML backend. It enables efficient dimensionality reduction, feature extraction, and pattern discovery in single-cell and spatial genomics data, with seamless integration into Bioconductor workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MicTott/BiocNMF}
  \item \url{https://mictott.github.io/BiocNMF/}
  \item Report bugs at \url{https://github.com/MicTott/BiocNMF/issues}
}

}
\author{
\strong{Maintainer}: Michael Totty \email{mictott@gmail.com}

}
\keyword{internal}
